+++ app/containers/ContactSearchResult/index.js
@@ -19,6 +19,7 @@ export class ContactSearchResult extends React.Component {
     };
 
     this.assignContact = this.assignContact.bind(this);
+    this.editContact = this.editContact.bind(this);
   }
 
   styles = {
@@ -60,11 +61,22 @@ export class ContactSearchResult extends React.Component {
   assignContact() {
     this.props.assignContact(this.props.contact);
   }
+  editContact() {
+    this.props.editContact(this.props.contact);
+  }
 
   render() {
     return (
       <div style={[this.styles.base, this.props.style]}>
-        <Contact contact={this.props.contact} style={[this.styles.contact]} showCompactView={!this.state.expanded} assign={this.assignContact} showControls={!this.props.loading}></Contact>
+        <Contact
+          contact={this.props.contact}
+          style={[this.styles.contact]}
+          showCompactView={!this.state.expanded}
+          isAssigned={this.props.isAssigned}
+          assign={this.assignContact}
+          edit={this.editContact}
+          showControls={!this.props.loading}
+        />
         <div onClick={() => this.setState({ expanded: !this.state.expanded })} style={this.styles.expandToggle} >
           <span style={this.styles.ellip}>...</span>
         </div>
@@ -77,7 +89,13 @@ ContactSearchResult.propTypes = {
   style: React.PropTypes.object,
   contact: React.PropTypes.object.isRequired,
   assignContact: React.PropTypes.func.isRequired,
+  isAssigned: React.PropTypes.bool,
+  editContact: React.PropTypes.func.isRequired,
   loading: React.PropTypes.bool,
+};
+
+ContactSearchResult.defaultProps = {
+  isAssigned: false,
 };
 
 export default Radium(ContactSearchResult);