+++ app/containers/ContactsControl/index.js
@@ -15,6 +15,7 @@ import selectContactsControl, { selectSelectedInteraction, selectAttributes } fr
 import { setSearchResults, clearSearchResults } from './actions';
 import { assignContact } from '../AgentDesktop/actions';
 
+import IconSVG from 'components/IconSVG';
 import Button from 'components/Button';
 import Filter from 'components/Filter';
 import Icon from 'components/Icon';
@@ -33,6 +34,7 @@ export class ContactsControl extends React.Component {
       query: this.expandQuery(this.props.selectedInteraction.query, this.props.attributes),
       loading: false,
       isEditing: false,
+      unassignedContactEditing: {},
     };
 
     this.setSearching = this.setSearching.bind(this);
@@ -41,7 +43,9 @@ export class ContactsControl extends React.Component {
     this.searchContacts = this.searchContacts.bind(this);
     this.renderResults = this.renderResults.bind(this);
     this.renderContactView = this.renderContactView.bind(this);
-    this.setEditing = this.setEditing.bind(this);
+    this.editUnassignedContact = this.editUnassignedContact.bind(this);
+    this.editAssignedContact = this.editAssignedContact.bind(this);
+    this.newContact = this.newContact.bind(this);
     this.getBannerHeader = this.getBannerHeader.bind(this);
     this.getViewControlHeader = this.getViewControlHeader.bind(this);
     this.getSearchControlHeader = this.getSearchControlHeader.bind(this);
@@ -105,8 +109,16 @@ export class ContactsControl extends React.Component {
     this.props.setContactAction(this.props.selectedInteraction.interactionId, 'search');
   }
 
-  setEditing() {
-    this.setState({ isEditing: true });
+  editUnassignedContact(contact) {
+    this.setState({ isEditing: true, unassignedContactEditing: contact });
+  }
+
+  newContact() {
+    this.setState({ isEditing: true, unassignedContactEditing: {} });
+  }
+
+  editAssignedContact() {
+    this.setState({ isEditing: true, unassignedContactEditing: {} });
   }
 
   setViewing() {
@@ -121,7 +133,7 @@ export class ContactsControl extends React.Component {
   }
 
   setNotEditing() {
-    this.setState({ isEditing: false });
+    this.setState({ isEditing: false, unassignedContactEditing: {} });
   }
 
   getSearchControlHeader() {
@@ -153,7 +165,7 @@ export class ContactsControl extends React.Component {
     return (
       <div style={this.styles.controlHeader}>
         <div style={this.styles.buttonSet}>
-          <Button id="contact-edit-btn" style={this.styles.leftButton} onClick={this.setEditing} text={messages.edit} type="secondary" />
+          <Button id="contact-edit-btn" style={this.styles.leftButton} onClick={this.editAssignedContact} text={messages.edit} type="secondary" />
           <Button id="contact-search-btn" style={this.styles.rightButton} onClick={this.setSearching} iconName="search" type="secondary" />
         </div>
       </div>
@@ -175,13 +187,17 @@ export class ContactsControl extends React.Component {
     switch (this.props.selectedInteraction.contactAction) {
       case 'view':
         if (this.state.isEditing) {
-          return this.getBannerHeader(<FormattedMessage {...messages.editingBanner} />);
+          return this.getBannerHeader(<FormattedMessage {...messages.contactEditingBanner} />);
         }
         return this.getViewControlHeader();
       case 'search':
       default:
         if (this.state.isEditing) {
-          return this.getBannerHeader(<FormattedMessage {...messages.newContactBanner} />);
+          if (Object.keys(this.state.unassignedContactEditing).length === 0) {
+            return this.getBannerHeader(<FormattedMessage {...messages.newContactBanner} />);
+          } else {
+            return this.getBannerHeader(<FormattedMessage {...messages.contactEditingBanner} />);
+          }
         }
         return this.getSearchControlHeader();
     }
@@ -194,7 +210,13 @@ export class ContactsControl extends React.Component {
       case 'search':
       default:
         if (this.state.isEditing) {
-          return <Contact save={this.handleSave} cancel={this.setNotEditing} style={this.styles.mainContact} isEditing={this.state.isEditing} />;
+          return (<Contact
+            save={this.handleSave}
+            cancel={this.setNotEditing}
+            style={this.styles.mainContact}
+            isEditing={this.state.isEditing}
+            contact={this.state.unassignedContactEditing}
+          />);
         }
         return this.renderResults();
     }
@@ -315,6 +337,13 @@ export class ContactsControl extends React.Component {
     leftGutter: {
       width: '52px',
     },
+    loading: {
+      display: 'flex',
+      justifyContent: 'center',
+    },
+    loadingIcon: {
+      height: '60px',
+    },
   };
 
   saveCallback(error, topic, response) {
@@ -323,6 +352,7 @@ export class ContactsControl extends React.Component {
     if (error) {
       console.error(error); // TODO: proper error display to user
     } else {
+      this.props.clearSearchResults();
       this.setNotEditing();
     }
   }
@@ -338,25 +368,43 @@ export class ContactsControl extends React.Component {
 
   assignContactToSelected(contact) {
     this.props.assignContact(this.props.selectedInteraction.interactionId, contact);
-    this.setViewing();
+  }
+
+  getLoader() {
+    return (
+      <div id="loadingContainer" style={ this.styles.loading }>
+        <IconSVG style={this.loadingIcon} id="loadingIcon" name="loading" />
+      </div>
+    );
   }
 
   renderResults() {
-    let results;
+    const results = [];
     if (this.props.selectedInteraction.query && Object.keys(this.props.selectedInteraction.query).length) {
-      const resultsMapped = this.props.results.map((contact) => <ContactSearchResult style={this.styles.contactResult} key={contact.id} contact={contact} assignContact={this.assignContactToSelected} loading={this.state.loading} />);
-      results = (
+      const resultsMapped = this.props.results.map(
+        (contact) =>
+          <ContactSearchResult
+            style={this.styles.contactResult}
+            key={contact.id}
+            isAssigned={this.props.selectedInteraction && this.props.selectedInteraction.contact && this.props.selectedInteraction.contact.id === contact.id}
+            contact={contact}
+            assignContact={this.assignContactToSelected}
+            editContact={this.editUnassignedContact}
+            loading={this.state.loading}
+          />);
+      results.push(
         <InfiniteScroll
           loadMore={this.searchContacts}
           hasMore={this.props.resultsCount === -1 || this.props.results.length < this.props.resultsCount}
-          loader={<div className="loader"><FormattedMessage {...messages.loading} /></div>}
+          loader={this.getLoader()}
           useWindow={false}
         >
           { resultsMapped }
         </InfiniteScroll>
       );
-    } else {
-      results = (
+    }
+    if (this.props.resultsCount < 1 && !this.state.loading) {
+      results.push(
         <div id="results-placeholder" style={this.styles.resultsPlaceholder}>
           <div style={this.styles.resultsPlaceholderTitle}>
             <Icon name="search" />
@@ -370,7 +418,7 @@ export class ContactsControl extends React.Component {
           <div style={{ margin: '5px 0' }}>
             <FormattedMessage {...messages.or} />
           </div>
-          <Button id="createNewRecord" type="secondary" text="Create New Record" onClick={this.setEditing}></Button>
+          <Button id="createNewRecord" type="secondary" text="Create New Record" onClick={this.newContact}></Button>
         </div>
       );
     }
@@ -379,7 +427,7 @@ export class ContactsControl extends React.Component {
 
   renderContactView() {
     return Object.keys(this.props.selectedInteraction.contact).length ?
-      <Contact style={this.styles.mainContact} contact={this.props.selectedInteraction.contact} loading={this.state.loading} save={this.handleSave} cancel={this.setNotEditing} isEditing={this.state.isEditing} />
+      <Contact style={this.styles.mainContact} isAssigned contact={this.props.selectedInteraction.contact} loading={this.state.loading} save={this.handleSave} cancel={this.setNotEditing} isEditing={this.state.isEditing} />
       :
       ''; // TODO: loading animation
   }